requirements.txt
Flask==3.0.0
pyserial==3.5
qrcode==7.4.2
Pillow==10.1.0
